requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Deprecated==1.2.13
discord.py==1.7.3
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
mathlibtools==1.0.0
multidict==6.0.2
networkx==2.6.2
pycparser==2.20
pydot==1.4.2
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
smmap==4.0.0
toml==0.10.2
tqdm==4.62.2
typing_extensions==4.3.0
urllib3==1.26.6
wrapt==1.12.1
yarl==1.7.2
